+++ lib/crimson.dart
@@ -2,4 +2,3 @@ library crimson;
 
 export 'src/annotations.dart' hide JsonIgnore, JsonKebabCase, JsonSnakeCase;
 export 'src/crimson.dart';
-export 'src/json_converter.dart';